Models/requirements.txt
albumentations==1.4.18
cmapy==0.6.6
matplotlib==3.5.3
numpy==2.2.5
onnx==1.17.0
onnxruntime==1.21.0
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.11.0.86
Pillow==11.3.0
pytorch_model_summary==0.1.2
thop==0.1.1.post2209072238
torch==2.7.0
torchvision==0.22.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
